move buffer code to common
